Add URL entry boxes for firmament URL and OnClick button URL
Add firmament URL textbox as a global setting. This allows for the user to update the URL in case it changes in the future.

Add OnClick button URL, allows the user to pick what URL is launched when the button is pressed.

Fixed issue where on initial loading of the plugin, the menu is not populated due to PI acquring global settings before Plugin generated them. PI now reloads UI elements when given a command through sendToPropertyInspector.

Some code refactoring