Add preset image selection feature
Add feature to allow for a few preset images to be selected in the PI to be used.

Fixed issue with PI function insertDropDownOption for when optGroupID is null.

Set Button URL to have a default value pointing to the Builders' Progress Report page.

Fixed issue where menu was inserted twice sometimes on load. Fixed by having the menu clear itself when reloading from global settings.

Add images to Resorces and Icons.

Update Readme